v1.3.1.2 update

+++ src/bitcoinrpc.h
@@ -126,7 +126,7 @@ extern const CRPCTable tableRPC;
 
 extern int64_t nWalletUnlockTime;
 extern int64_t AmountFromValue(const json_spirit::Value& value);
-extern json_spirit::Value ValueFromAmount(int64_t amount);
+extern json_spirit::Value ValueFromAmount(__int128 amount);
 extern json_spirit::Value ValueFromBigNumAmount(CBigNum amount);
 extern double GetDifficulty(const CBlockIndex* blockindex = NULL);
